Extract drop_null_fields helpers into a shared library
Move the row-grouping and null-pattern logic out of the operator's
translation unit into a public `tenzir::drop_null_fields` function so a
matching function plugin can reuse the same implementation. The operator
becomes a thin wrapper, leaving its observable behavior unchanged.
